+++ common/src/main/java/net/dodogang/plume/block/PlumeStairsBlock.java
@@ -1,0 +1,10 @@
+package net.dodogang.plume.block;
+
+import net.minecraft.block.BlockState;
+import net.minecraft.block.StairsBlock;
+
+public class PlumeStairsBlock extends StairsBlock {
+    public PlumeStairsBlock(BlockState blockState, Settings settings) {
+        super(blockState, settings);
+    }
+}